+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+### Added
+
+- Ability to use enums to reference services and specify the returned object type
+
 ### Removed
 
 - Support for PHP `8.1`
+
+### Deprecated
+
+- Using `string`s as a service name